+++ src/data.ts
@@ -231,7 +231,33 @@ let structs_build = new ProblemSet(
 function pick_random_el(arr: any[]) {
   return arr[Math.floor(Math.random() * arr.length)];
 }
-let all: ProblemSet[] = [racket_math_expressions, infix_to_prefix, racket_string_practice, ran_var_arith, build_a_string, comparisons, logical, func_scope_eval, func_scope_eval_2, structs_build];
+
+// files and their names stored at : https://github.com/jestarray/howtocode/starter
+function exercise_from_filenames() {
+  // TODO: make this an object with youtube links
+  let filenames = [
+    "4-string-first_starter.rkt",
+    "4.1-string-last_starter.rkt",
+    "4.2-string-rest_starter.rkt",
+    "4.3-string-remove-last_starter.rkt",
+    "4.4-formalize_starter.rkt",
+    "5-image-area_starter.rkt",
+    "5.1-red-x_starter.rkt",
+    "5.2-cancel_starter.rkt",
+    "6-ensure-question_starter.rkt",
+    "6.1-punctuate_starter.rkt",
+  ];
+  return filenames.map((item) => {
+    let dash_index = item.indexOf("-");
+    let problem_number = parseFloat(item.slice(0, dash_index));
+    let problem_name = item.slice(dash_index + 1).replace("_starter.rkt", "");
+    return new ProblemSet(problem_name, problem_number, undefined, [new Problem("submission", 
+    `<h1>⬇️<a href="https://howtocode.pages.dev/starter/${item}" download>Download the starter file</a></h1>`, "")], 
+    [{url_title: "Solution", url: `https://howtocode.pages.dev/solutions/${item.replace("_starter.rkt", "_solution.rkt")}`, additional: ""}])
+  });
+}
+
+let all: ProblemSet[] = [racket_math_expressions, infix_to_prefix, racket_string_practice, ran_var_arith, build_a_string, comparisons, logical, func_scope_eval, func_scope_eval_2, structs_build, ...exercise_from_filenames()];
 //WARNING: do not stringify this! we need the gen function and a copy of the original questions for resetting and other things
 export const TOC_original: ProblemSet[] = all.map((val) => Object.freeze(val));